Rename: dashboard 그룹라우팅
