+++ packages/web/src/app/(dashboard)/chat/layout.tsx
