+++ packages/web/src/app/(dashboard)/layout.tsx
