+++ packages/web/src/components/Friends/friends.module.css
@@ -1,0 +1,14 @@
+.ul {
+  max-height: 74px;
+  display: flex;
+  gap: 16px;
+  align-content: flex-start;
+  justify-content: flex-start;
+
+  overflow-x: scroll;
+  overflow-y: hidden;
+}
+
+.ul::-webkit-scrollbar {
+  display: none;
+}
